Refactoring user profile component into multiple pieces - resource.ts cahnges as well
